requirements.txt
streamlit
rdkit
streamlit-ketcher
pillow
py3Dmol
